ReconfigureAsyns: FireAndForget for subscription re-establish
This is effectively the behavior we had before (but we are ensured connected now). We just want the count here and to start them, so let's do that also pipelining the `n` commands as we did before instead of awaiting each. We also don't need the `Task` overhead. This makes moving to a variadic command set for this less urgent.