v2.5 work: BacklogPolicy (#1912)
This is a work in progress of the Backlog Policy bits to help address #1864. The overall change here is to make the message pathway backlog commands when an endpoint is down. It does not hand off to another viable endpoint in this iteration (future plans there). If the policy is set to `FailFast`, we should get the old behavior not queueing anything. In the new path:
- Handshake commands go straight to the IO pipe (the only way this would fail is if we lost a pipe _during a handshake_, in which case things are ill and let's backoff until the next heartbeat anyway).
- If connected and there is no backlog queue, commands go into the pipe
- If disconnected or there is a backlog queue, non-internal and non-handshake commands go into the backlog, so that they are played to the pipe in order

TODO List:
- [x] Implement simplified backlog queue (3 queues -> 1, no handoffs)
- [x] Implement backlog policy and config
- [x] Add initial testing
- [x] Find intermittent async hang in local test runner (test: `QueuesAndFlushesAfterReconnecting`) - this may be boneheaded or a fundamental flaw
- [x] Add test for sync commands (e.g. `.Ping()`) and that they recover appropriately

Co-authored-by: mgravell <marc.gravell@gmail.com>