Client name: append library version suffix (#1985)
This changes the default client name to append "(v2.5.x)" on the end (no spaces because Redis doesn't support spaces in client names). It won't append if someone is setting their client name explicitly in options, only for our default, so should be a safe change.

I'm not sure if a `Utils` class (internal) is best here, but seemed like the right path. I really didn't want to ref version from `ExceptionFactory` in the multiplexer...