Configuration: An idea on separation (#1987)
This is just an idea pitch (scrutinize heavily!) to see if the direction even works. The intent is for a consumer or wrapper library to specify a default options provider (or extend the built-in ones). For example Azure could override `IsMatch` and maintain its own domain list, as could AWS, etc.

Problems with this:
- Modifying the `KnownProviders` list on `DefaultOptionsProvider` is awkward, but `ConfigurationOptions` can also get a `.Defaults` set directly, or...other ideas? Maybe `GetForEndpoints` should get on `ConfigurationOptions` as a `public static` instead?

Eyes should be given here on new public members (and naming!) on `ConnectionMultiplexer`. Maintenance events are meant to be extensible as part of this change.