Tests: normalize all the things (#2097)
In recent PRs, through no fault of their own, I can see a lot of inconsistency because they're based on inconsistency in the existing test suite. This is one major overhaul to unify how things are laid out, connection naming, etc.

Once this builds, I'll merge in then tidy up PRs in progress to be on the common scheme and have less confusion for contributors going forward.