Support the XAUTOCLAIM command. (#2095)
Added [XAUTOCLAIM](https://redis.io/commands/xautoclaim/) support as part of #2055 

The XCLAIM command has two methods, StreamClaim & StreamClaimIdsOnly. Since the XAUTOCLAIM result is a bit more complex than the XCLAIM command, I opted to consolidate to a single method and single result type. 

To return just the message IDs in the result, the `idsOnly` parameter should be set to `true`. The caveat here is that the `StreamEntry` instances will only have `Id` populated when `idsOnly` is `true`, the `Values` array will be empty. This behavior is called out in the XML docs for the method. 

I wasn't sure if this is the proper "feel" you want for this command or not. Let me know if you want to break it up like `StreamClaim`.

Co-authored-by: Nick Craver <nrcraver@gmail.com>