MOVED: Proactively reconfigure when we hit a MOVED response
I'm not 100% sure about this because if there is a MOVED happening (e.g. bad proxy somewhere) this would just continually re-run...but only once every 5 seconds. Overall though, we linger in a bad state retrying moves until a discovery happens today and this could be resolved much faster.

Meant to help address #1520, #1660, #2074, and #2020.