Timeouts: better exceptions when messages timeout from the backlog
This is to better inform users when timeouts are happening in the backlog buffer. However, it's also exposed a race in the test suite where alone on in a small batch tests run fine. When we run everything together, then `DisconnectAndNoReconnectThrowsConnectionExceptionAsync` will hang indefinitely. But of course not debugging, we have some sort of fun race here.

Note the new handling inside ProcessBacklogAsync which I observed trying to figure this out - we have 1 test that's in a disposal race with backlogged items and we need to make sure everything in there gets punted out to avoid async hangs forever. But there's still another cause of this somewhere.