Exceptions: include Timer.ActiveCount in .NET Core stats (#2500)
We're seeing some instances of quite delayed timeouts and at least in two deeper investigations it was due to a huge number of timers firing and the backlog timeout assessment timer not triggering for an extended period of time. To help users diagnose this, adding the cheap counter to the .NET Core pool stats where `Timer.ActiveCount` is readily available.

This is available in all the .NET (non-Framework) versions we support.

Before:
> ...POOL: (Threads=25,QueuedItems=0,CompletedItems=1066)

After:
> ...POOL: (Threads=25,QueuedItems=0,CompletedItems=1066,Timers=46)

See #2477 for another possible instance of this.