CheckTrustedIssuer: Fixes for invalid chains
This issue was brought to my attention last night (thanks reporter!): https://github.com/dotnet/dotnet-api-docs/pull/6660

This changeset ensures that we do not honor self-signed certs or partial/broken chains as a result of `X509VerificationFlags.AllowUnknownCertificateAuthority` downstream and adds a few tests and utilities to generate test certificates (currently valid for ~9000 days). Instead we are checking that the certificate we're being told to trust is explicitly in the chain, given that the result of `.Build()` cannot be trusted for this case.