Build: Simplify Docker layout, fix Envoy, and upgrade tests (#2774)
It seems the Envoy apts have gone missing breaking out build, so instead of relying on the install let's docker compose their image as a proxy against the Redis supervisor instance as a simpler and faster-to-start setup that also works.

This rearranged some things to simplify the Docker story overall. A move to AzDO or just GitHub builds would simplify everything further, but we need to figure out Windows testing against a Docker setup in CI.

Note: we still can't use Linux containers on a Windows GitHub Actions host (https://github.com/actions/runner/issues/904), so this remains much more complicated and not-really-testing-the-real-thing in the Windows front.